backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic[email]==2.6.3
email-validator==2.1.1
